move load more logic into the adapter
The adapter uses the position of the view in the
onBindViewHolder method to calculate when to trigger
the load more callback. No more onScrollListener

After moving this logic from the custom RecyclerView,
it became pretty much useless. So it has been completely
removed, in favor of just extending the adapter, and using
it with any recycler view

Breaking changes:

    - InfiniteRecyclerView is gone. Using the normal
      RecylerView, or any other extension of it, in your
      layout, should work fine.
    - All the methods previously called on the RecyclerView
      in the Activity/Fragment, should now be called on the
      adapter
    - Removed deprecated moreDataLoaded()